Move additional signature traits to core
