+++ identity-core/src/crypto/mod.rs
@@ -12,6 +12,7 @@ pub use self::key::KeyPair;
 pub use self::key::PublicKey;
 pub use self::key::SecretKey;
 pub use self::proof::JcsEd25519Signature2020;
+pub use self::signature::SetSignature;
 pub use self::signature::SigName;
 pub use self::signature::SigSign;
 pub use self::signature::SigVerify;
@@ -19,3 +20,5 @@ pub use self::signature::Signature;
 pub use self::signature::SignatureData;
 pub use self::signature::SignatureOptions;
 pub use self::signature::SignatureValue;
+pub use self::signature::TrySignature;
+pub use self::signature::TrySignatureMut;
