+++ identity-did/src/verifiable/mod.rs
@@ -10,6 +10,3 @@ pub use self::document::Document;
 pub use self::ld_suite::LdSuite;
 pub use self::properties::Properties;
 pub use self::traits::ResolveMethod;
-pub use self::traits::SetSignature;
-pub use self::traits::TrySignature;
-pub use self::traits::TrySignatureMut;
