+++ identity-did/src/verifiable/traits.rs
@@ -1,74 +1,10 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use identity_core::crypto::Signature;
-
 use crate::error::Error;
 use crate::error::Result;
 use crate::verification::MethodQuery;
 use crate::verification::MethodWrap;
-
-pub trait TrySignature {
-  fn signature(&self) -> Option<&Signature>;
-
-  fn try_signature(&self) -> Result<&Signature> {
-    self.signature().ok_or(Error::QuerySignatureNotFound)
-  }
-}
-
-impl<'a, T> TrySignature for &'a T
-where
-  T: TrySignature,
-{
-  fn signature(&self) -> Option<&Signature> {
-    (**self).signature()
-  }
-}
-
-impl<'a, T> TrySignature for &'a mut T
-where
-  T: TrySignature,
-{
-  fn signature(&self) -> Option<&Signature> {
-    (**self).signature()
-  }
-}
-
-// =============================================================================
-// =============================================================================
-
-pub trait TrySignatureMut: TrySignature {
-  fn signature_mut(&mut self) -> Option<&mut Signature>;
-
-  fn try_signature_mut(&mut self) -> Result<&mut Signature> {
-    self.signature_mut().ok_or(Error::QuerySignatureNotFound)
-  }
-}
-
-impl<'a, T> TrySignatureMut for &'a mut T
-where
-  T: TrySignatureMut,
-{
-  fn signature_mut(&mut self) -> Option<&mut Signature> {
-    (**self).signature_mut()
-  }
-}
-
-// =============================================================================
-// =============================================================================
-
-pub trait SetSignature: TrySignatureMut {
-  fn set_signature(&mut self, signature: Signature);
-}
-
-impl<'a, T> SetSignature for &'a mut T
-where
-  T: SetSignature,
-{
-  fn set_signature(&mut self, signature: Signature) {
-    (**self).set_signature(signature);
-  }
-}
 
 // =============================================================================
 // =============================================================================
