+++ identity-iota/src/did/document.rs
@@ -15,16 +15,16 @@ use identity_core::convert::SerdeInto;
 use identity_core::crypto::JcsEd25519Signature2020;
 use identity_core::crypto::KeyPair;
 use identity_core::crypto::SecretKey;
+use identity_core::crypto::SetSignature;
 use identity_core::crypto::Signature;
 use identity_core::crypto::SignatureOptions;
+use identity_core::crypto::TrySignature;
+use identity_core::crypto::TrySignatureMut;
 use identity_did::did::DID;
 use identity_did::document::Document;
 use identity_did::verifiable::Document as VerifiableDocument;
 use identity_did::verifiable::LdSuite;
 use identity_did::verifiable::ResolveMethod;
-use identity_did::verifiable::SetSignature;
-use identity_did::verifiable::TrySignature;
-use identity_did::verifiable::TrySignatureMut;
 use identity_did::verification::MethodQuery;
 use identity_did::verification::MethodScope;
 use identity_did::verification::MethodType;
